generator: Check contained in hash
directly instead of explicitly requesting only the keys as a list and
converting that list to another list.
Checking directly for an element to be contained in a hash is much more
efficiently as this is done using hashing O(1) instead of waling the
list in half on average O(n).

Signed-off-by: Philipp Hahn <hahn@univention.de>